# Simplify pair computation on AEV