Because nan < number is always false.

Benchmark

Before
```
torchani.aev.cutoff_cosine - 0.1s
torchani.aev.radial_terms - 0.2s
torchani.aev.angular_terms - 0.6s
torchani.aev.compute_shifts - 0.0s
torchani.aev.neighbor_pairs - 0.0s
torchani.aev.neighbor_pairs_nopbc - 1.9s
torchani.aev.triu_index - 0.0s
torchani.aev.cumsum_from_zero - 0.1s
torchani.aev.triple_by_molecule - 2.5s
torchani.aev.compute_aev - 6.6s
Total AEV - 6.6s
NN - 13.0s
Epoch time - 29.6s
```

vs

After
```
torchani.aev.cutoff_cosine - 0.1s
torchani.aev.radial_terms - 0.2s
torchani.aev.angular_terms - 0.6s
torchani.aev.compute_shifts - 0.0s
torchani.aev.neighbor_pairs - 0.0s
torchani.aev.neighbor_pairs_nopbc - 1.7s
torchani.aev.triu_index - 0.0s
torchani.aev.cumsum_from_zero - 0.1s
torchani.aev.triple_by_molecule - 2.4s
torchani.aev.compute_aev - 6.3s
Total AEV - 6.3s
NN - 12.9s
Epoch time - 29.4s
```